Adding the Score hessian calls to SiloService. Fixing Scoring issues that popped up. Adding a ScoreTest that runs a trip and waits for the next agg cycle to run so it can compare scores.

+++ emulation/src/main/java/com/inthinc/device/resources/DeviceStatistics.java
@@ -22,7 +22,7 @@ public class DeviceStatistics {
         stop.setDate(System.currentTimeMillis());
         if (stop.getDelta(last) > 15000){
         	Log.i("Sent: %d, Time: %d, NotesPerMinute: %d",
-                    hessianCalls, getTimeDelta(), getCallsPerMinute());
+                    hessianCalls, getTimeDelta().intValue(), getCallsPerMinute());
             last.setDate(stop);
         }
         return true;
